refactor(breadcrumb): PR amends
