+++ components/breadcrumb/package.json
@@ -18,6 +18,7 @@
 	"dependencies": {
 		"@westpac/accessibility-helpers": "^0.0.1",
 		"@westpac/core": "^0.0.1",
+		"@westpac/icon": "^0.0.1",
 		"prop-types": "^15.6.2"
 	},
 	"peerDependencies": {
